changed wear package name an applicationID
