+++ wear/build.gradle
@@ -6,7 +6,7 @@ android {
     buildToolsVersion "22.0.1"
 
     defaultConfig {
-        applicationId "com.dexdrip.stephenblack.nightwatch"
+        applicationId "com.eveningoutpost.dexdrip"
         minSdkVersion 20
         targetSdkVersion 23
         versionCode 1
