+++ wear/src/main/java/com/eveningoutpost/dexdrip/LargeHome.java
@@ -1,10 +1,8 @@
-package com.dexdrip.stephenblack.dexdrip;
+package com.eveningoutpost.dexdrip;
 
-import android.content.Context;
 import android.graphics.Color;
 import android.view.LayoutInflater;
 
-import com.eveningoutpost.dexdrip.R;
 import com.ustwo.clockwise.WatchMode;
 
 public class LargeHome extends BaseWatchFace {
@@ -12,7 +10,7 @@ public class LargeHome extends BaseWatchFace {
     @Override
     public void onCreate() {
         super.onCreate();
-        LayoutInflater inflater = (LayoutInflater) getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+        LayoutInflater inflater = (LayoutInflater) getSystemService(LAYOUT_INFLATER_SERVICE);
         layoutView = inflater.inflate(R.layout.activity_home_large, null);
         performViewSetup();
     }
